pyproject.toml
[project]
name = "settings_setter"
version = "0.1.0"
description = "Aerobits settings setter"
readme = "README.md" 
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.11.1",
    "ruff>=0.8.0",
]
